Add onTap functionality to notification icons
Added an optional `onTap` parameter to `CustomNotifactionIconWithoutAlert` and `CustomNotifactionIconWithAlert` widgets to allow for tap events. Updated the `CustomHomeViewAppBar` widget to pass an empty `onTap` function to the `CustomNotifactionIconWithAlert` widget. This change enables interactive behavior for the notification icons, allowing for future implementation of tap actions such as opening a notification panel or navigating to a notifications view.